rhbz#1052142: partition_test fails with huge values in *space* system data
Introduced by:
9601ce1032fca9893b8e04d12f6bbab2c9f86729

+++ src/OVAL/results/oval_cmp_basic.c
@@ -51,7 +51,7 @@ oval_result_t oval_boolean_cmp(const bool state, const bool syschar, oval_operat
 	return OVAL_RESULT_ERROR;
 }
 
-oval_result_t oval_int_cmp(const int state, const int syschar, oval_operation_t operation)
+oval_result_t oval_int_cmp(const intmax_t state, const intmax_t syschar, oval_operation_t operation)
 {
 	switch (operation) {
 	case OVAL_OPERATION_EQUALS:
